extract shared test utils
